refactor(trusty-common): split kg.rs into submodules under 500-SLOC cap (#607)
Split the 815-SLOC store/kg.rs into 6 focused files under store/kg/:
- types.rs (56 SLOC): UndirectedSnapshot, KgEdge, Triple
- adjacency.rs (134 SLOC): private Adjacency struct + hydrate_adjacency
- graph.rs (370 SLOC): KnowledgeGraph struct, open, assert, retract, graph traversal
- ops.rs (212 SLOC): storage delegation methods (query, list, drawers, cascade delete)
- tests.rs (290 SLOC, test file): all inline tests from original kg.rs
- mod.rs (22 SLOC): thin re-export facade

Also fixes omitted deletion of dream.rs from commit 78a328ae (the dream/
directory was created but the original file was not deleted).

Fixes in retrieval split (companion fixes to 4bbd142d):
- Add missing imports to retrieval/tests.rs (PalaceId, Drawer, DrawerType,
  RoomType, Arc, RecallLog, HashMap, Uuid, VectorStore, uuid_prefix_eq,
  L1_NO_SIMILARITY_PENALTY)
- Make layers::L1_NO_SIMILARITY_PENALTY pub(super) for test access
- Fix redundant closure in handle.rs (clippy::redundant_closure)
- Add #[allow(clippy::module_inception)] to kg_redb/tests.rs

Updates .line-cap-allowlist.tsv to remove stale kg.rs entry.
