+++ .line-cap-allowlist.tsv
@@ -16,7 +16,6 @@ crates/trusty-analyze/src/lang/adapters/typescript.rs	560
 crates/trusty-analyze/src/main.rs	567
 crates/trusty-analyze/src/mcp/mod.rs	892
 crates/trusty-common/src/memory_core/retrieval/mod.rs	830
-crates/trusty-common/src/memory_core/store/kg.rs	815
 crates/trusty-common/src/monitor/search_tui.rs	1596
 crates/trusty-git-analytics/src/classify/mod.rs	793
 crates/trusty-git-analytics/src/classify/pipeline.rs	1247
